chore: merge coderetrx/impl/default to coderetrx/retrieval
